requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
pypdf
faiss-cpu
python-dotenv
streamlit
